reafctor to support a simpler logic
